+++ R/validate.R
@@ -79,6 +79,8 @@ validateNameArgument <- function(name,
 #' the individuals observation periods will be performed.
 #' @param checkAttributes Whether to check if attributes are present and
 #' populated correctly.
+#' @param checkPermanentTable Whether to check if the table has to be a
+#' permanent table.
 #' @param dropExtraColumns Whether to drop extra columns that are not the
 #' required ones.
 #' @param validation How to perform validation: "error", "warning".
@@ -118,6 +120,7 @@ validateCohortArgument <- function(cohort,
                                    checkMissingValues = FALSE,
                                    checkInObservation = FALSE,
                                    checkAttributes = FALSE,
+                                   checkPermanentTable = FALSE,
                                    dropExtraColumns = FALSE,
                                    validation = "error",
                                    call = parent.frame()) {
@@ -128,10 +131,11 @@ validateCohortArgument <- function(cohort,
   assertLogical(checkInObservation, length = 1)
   assertLogical(checkAttributes, length = 1)
   assertLogical(dropExtraColumns, length = 1)
+  assertLogical(checkPermanentTable, length = 1)
 
   assertClass(cohort, class = c("cohort_table", "cdm_table"), all = TRUE, call = call)
 
-  if (is.na(tableName(cohort))) {
+  if (is.na(tableName(cohort)) & checkPermanentTable) {
     missingCohortTableNameError(cdmReference(cohort), validation = validation)
   }
 
@@ -464,35 +468,39 @@ getWindowNames <- function(window, snakeCase) {
 
   if (isTRUE(snakeCase)) {
     if (is.null(windowNames)) {
-      windowNames <- lapply(window, getname)
+      windowNames <- purrr::map_chr(window, getname)
     } else {
-      windowNames[windowNames == ""] <-
-        lapply(window[windowNames == ""], getname)
+      id <- windowNames == ""
+      windowNames[id] <- purrr::map_chr(window[id], getname)
+      newNames <- toSnakeCase(windowNames)
+      differentNames <- which(windowNames != newNames)
+      if (length(differentNames) > 0) {
+        newName <- newNames[differentNames]
+        oldName <- windowNames[differentNames]
+        changes <- paste0("`", oldName, "` -> `", newName, "`") |>
+          rlang::set_names(rep("*", length(newName)))
+        cli::cli_inform(c("window names casted to snake_case: ", changes))
+      }
+      windowNames <- newNames
     }
   } else {
     if (is.null(windowNames)) {
-      windowNames <- lapply(window, getname2)
+      windowNames <- purrr::map_chr(window, getname2)
     } else {
-      windowNames[windowNames == ""] <-
-        lapply(window[windowNames == ""], getname2)
+      id <- windowNames == ""
+      windowNames[id] <- purrr::map_chr(window[id], getname2)
     }
   }
   windowNames
 }
 assertWindowName <- function(window, snakeCase, call = parent.frame()) {
   names(window) <- getWindowNames(window, snakeCase = snakeCase)
-  lower <- lapply(window, function(x) {
-    x[1]
-  }) |> unlist()
-  upper <- lapply(window, function(x) {
-    x[2]
-  }) |> unlist()
+  lower <- purrr::map_dbl(window, \(x) x[1])
+  upper <- purrr::map_dbl(window, \(x) x[2])
 
   if (any(lower > upper)) {
-    cli::cli_abort("First element in window must be smaller or equal to
-                   the second one",
-                   call = call
-    )
+    "First element in window must be smaller or equal to the second one" |>
+      cli::cli_abort(call = call)
   }
   if (any(is.infinite(lower) & lower == upper & sign(upper) == 1)) {
     cli::cli_abort("Not both elements in the window can be +Inf", call = call)
@@ -1067,6 +1075,9 @@ validateNameStyle <- function(nameStyle,
 validateStrataArgument <- function(strata,
                                    table,
                                    call = parent.frame()) {
+  if (is.character(strata)) {
+    strata <- list(strata)
+  }
   assertList(strata, class = "character", call = call)
   cols <- colnames(table)